🎨 modularize note
